I modified the folder structure so the controllers could display the correct view, added some more views, and learned how to use httpContext.Session: I added the Session and memory cache to the builder. I'm still in uncharted territory. --->(attempted to push on 11/30/23)
I created a way for the user to log in to their account. I added login validation and, if valid, stored the user name in the Session to change the navigation bar to include the user's user name and a Logout button.

I also started working on the Add Item I action.